+++ client/src/App.tsx
@@ -3,8 +3,8 @@ import HelloWeb3 from "./components/pages/HelloWeb3";
 import SignUp from "./components/pages/SignUp";
 import SignIn from "./components/pages/SignIn";
 import Mypage from "./components/pages/Mypage";
-import AuthProvider from "./components/contexts/AuthContext";
-import AuthRequired from "./components/organisms/AuthRequired";
+import OptionalAuthProvider from "./components/organisms/OptionalAuthProvider";
+import AuthProvider from "./components/organisms/AuthProvider";
 import ImageUpload from "./components/pages/ImageUpload";
 import Header from "./components/organisms/Header";
 import NotFound from "./components/pages/NofFound";
@@ -12,18 +12,18 @@ import NotFound from "./components/pages/NofFound";
 const App = () => {
   return (
     <>
-      <AuthProvider>
+      <OptionalAuthProvider>
         <Header />
-      </AuthProvider>
+      </OptionalAuthProvider>
       <hr />
       <Routes>
         <Route path="/" element={<HelloWeb3 />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/signin" element={<SignIn />} />
-        <Route path="/mypage" element={<AuthRequired element={<Mypage />} />} />
+        <Route path="/mypage" element={<AuthProvider element={<Mypage />} />} />
         <Route
           path="/image-upload"
-          element={<AuthRequired element={<ImageUpload />} />}
+          element={<AuthProvider element={<ImageUpload />} />}
         />
         <Route path="*" element={<NotFound />} />
       </Routes>
